+++ apps/ranger-next/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ranger-theme/nextjs
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @ranger-theme/nextjs-config@0.3.0
+  - @ranger-theme/pwa-config@0.1.0
+
 ## 1.1.1
 
 ### Patch Changes